Move transformation scenario views out into their own file
