+++ core/views/__init__.py
@@ -4,3 +4,4 @@ from .organization import *
 from .record_group import *
 from .job import *
 from .record import *
+from .transformation_scenario import *
